handle exception for user

+++ src/services/user.ts
@@ -5,6 +5,8 @@ import bcrypt from "bcrypt";
 import loggerWithNameSpace from "../utils/logger";
 import { UUID } from "crypto";
 import config from "../config";
+import { BadRequestError } from "../error/BadRequestError";
+import { NotFoundError } from "../error/NotFoundError";
 
 const logger = loggerWithNameSpace("service: user");
 
@@ -13,9 +15,14 @@ export async function createUser(user: IUser) {
 
   let newUser: IUser;
 
-  const hashSaltValue = 10;
+  //to prevent multiple user with same email
+  if ((await getUserByEmail(user.email)).length !== 0) {
+    logger.error(`Email is already used:${user.email}`);
 
-  const password = await bcrypt.hash(user.password, hashSaltValue);
+    throw new BadRequestError("Email is already used");
+  }
+
+  const password = await bcrypt.hash(user.password, config.bcryptSalt);
 
   newUser = {
     ...user,
@@ -40,6 +47,12 @@ export function getUserByEmail(email: string) {
 export async function updateUser(id: UUID, user: IUser) {
   logger.info("Updating a user");
 
+  const existingUser = (await getUsers({ id: id }))[0];
+
+  if (!existingUser) {
+    throw new NotFoundError("user not found");
+  }
+
   const password = await bcrypt.hash(user.password, config.bcryptSalt);
   user.password = password;
 
@@ -48,6 +61,12 @@ export async function updateUser(id: UUID, user: IUser) {
 
 export async function deleteUser(id: UUID) {
   logger.info("Deleting a user");
+
+  const existingUser = (await getUsers({ id: id }))[0];
+
+  if (!existingUser) {
+    throw new NotFoundError("user not found");
+  }
 
   return UserModel.delete(id);
 }
